Smoothly refined mesh (#3)
* add UnstructuredGridFunc

* Update MixingLayer3D.h

* add trafo

* added transformation function

* added transformation function

* LS Advection tracker only on MPI rank 0

* LS Advection tracker only on MPI rank 0

* reverted: LS Advection tracker only on MPI rank 0

* transformation with y' = tan(y)

* transformation with parabola

* transformation with parabola

* added no-output-interval (n-no-out) option

* added random_u_cube_k048_half.txt

* moved random_u's and meshes into subdirectory

* added ncheckpoint option

* added ncheckpoint option - only if output is not turned off

* added random_u_cube_k0172_half.txt

* do not turn of sl stats

* scale domain in step-file

* fixed n-no-out

* propagating repetitions all the way down

* calculate deltas after transform()

* Update ShearLayerStats.cpp

ignoring nooutput interval

* Update CompressibleCFDSolver.h

ignore no-output in compressible output

* calculate deltas after transform() - and relative to dT0

* fixed first couple of warnings

* fixed no-output by using int instead of size_t

* fixed more warnings

* removed array initialization as 0.0 to remove missing brackets warnings

* replaced deprecated n_locally_owned_dofs_per_processor()

* fixed more warnings

* Do VTK every step if MaxP is >0.6

* fixed most warnings from omni

* Separate no stats (#2)

* Update SolverConfiguration.cpp

* Update SolverConfiguration.h

* Update ShearLayerStats.cpp

* Update ShearLayerStats.h

* Update step-mixingLayer.cpp

* get current git branch

* fix m_dUx to m_dU0 and use exact coordinates

* fixed nround - hopefully 10^-10 is small enough

* make nround depend on minimum dof distance, rounded to 4 decimal places

* make nround depend on minimum dof distance, rounded to 8 decimal places
print min and max dof distances

* added maxVertexDistance and maxDOFdistance

* check uniqueness of coordinates by rounding

* get vertex distances in all directions

* update mesh info output

* default ncoordsround to 10

* predict Deltas for ref-level 3

* update defaults

* correctly divide integrals over shearlayer thickness

* added integration test

* update integration method

* update integration method

* always set fi

* use glc elements per default

* update integration method

* update random_u

* update integration method

* update DOF calculation to glc

* fix integral over whole domain

* readded getMinimumDoFDistanceGLL for utilities test

* print outputdir

* reactivate turbulentchannel

* log boundary condition

* log number of cells and number of cells

* take len_x as whole length

* added equidistant points option

* update random velocities

* update LOG

* update random_u to p=5
use commandlineparser more

* do not build turbulentChannel

* fix stencil to Q45

* changed default grid trafo
added LOG outputs

* new velocity field with smoother frequency scaling

* update

* output m_dUx

* output m_dUx m_ReOmega

* use dU0 instead of dUx for calculations

* double width of randU scaling

* minor update to derivative calculation

* more test functions (TODO: do not normalize RMS, take root, and mean instead of sum)

* added ux_Re calculation

* added resolutions.ods

* correct randU scaling (2xdT, not 4x)

* use ux for deltaOmega

* use ux for deltaOmega

+++ src/test/stencils/D2Q19H_test.cpp
@@ -34,7 +34,7 @@ BOOST_AUTO_TEST_SUITE(D2Q19H_test)
         BOOST_CHECK_CLOSE(dqmodel.getSpeedOfSoundSquare(), 1./3.,0.000001);
         std::vector<double> weights = dqmodel.getWeights();
         double sum =0.0;
-        for(int i = 0; i < dqmodel.getQ(); i++)
+        for(size_t i = 0; i < dqmodel.getQ(); i++)
             sum+=weights[i];
         BOOST_CHECK_CLOSE(sum,1.0,0.0000001);
